+++ src/rs/Class7_Sub2_Sub1.java
@@ -22,15 +22,15 @@ public class Class7_Sub2_Sub1 extends Class7_Sub2 {
 			int i_0_ = -71;
 			return i_0_;
 		}
-		if ((double) JunkTex.aFloat2854 == 3.0) {
+		if (JunkTex.aFloat2854 == 3.0) {
 			int i_1_ = 37;
 			return i_1_;
 		}
-		if ((double) JunkTex.aFloat2854 == 4.0) {
+		if (JunkTex.aFloat2854 == 4.0) {
 			int i_2_ = 50;
 			return i_2_;
 		}
-		if (6.0 == (double) JunkTex.aFloat2854) {
+		if (6.0 == JunkTex.aFloat2854) {
 			int i_3_ = 75;
 			return i_3_;
 		}
