+++ src/rs/Class94.java
@@ -172,7 +172,7 @@ public class Class94 {
 			Class14_Sub2_Sub16 class14_sub2_sub16 = null;
 			return class14_sub2_sub16;
 		}
-		Class14_Sub2_Sub16 class14_sub2_sub16 = ((Class14_Sub2_Sub16) Class83.aClass52_1339.get((long) anInt1588));
+		Class14_Sub2_Sub16 class14_sub2_sub16 = ((Class14_Sub2_Sub16) Class83.aClass52_1339.get(anInt1588));
 		if (class14_sub2_sub16 != null) {
 			Class14_Sub2_Sub16 class14_sub2_sub16_0_ = class14_sub2_sub16;
 			return class14_sub2_sub16_0_;
@@ -182,7 +182,7 @@ public class Class94 {
 				Class134.aClass9_2598);
 		if (class14_sub2_sub16_2_ != null) {
 			class14_sub2_sub16_2_.method352(class148s, null);
-			Class83.aClass52_1339.put(class14_sub2_sub16_2_, (long) anInt1588);
+			Class83.aClass52_1339.put(class14_sub2_sub16_2_, anInt1588);
 		} else
 			aBoolean1554 = true;
 		Class14_Sub2_Sub16 class14_sub2_sub16_3_ = class14_sub2_sub16_2_;
@@ -609,7 +609,7 @@ public class Class94 {
 			Class14_Sub2_Sub19 class14_sub2_sub19 = null;
 			return class14_sub2_sub19;
 		}
-		long l = (((aBoolean1612 ? 1L : 0L) << 39) + (long) i_41_ - (-((aBoolean1577 ? 1L : 0L) << 35)
+		long l = (((aBoolean1612 ? 1L : 0L) << 39) + i_41_ - (-((aBoolean1577 ? 1L : 0L) << 35)
 				- ((long) anInt1597 << 36) + -((aBoolean1472 ? 1L : 0L) << 38) + -((long) anInt1570 << 40)));
 		Class14_Sub2_Sub19 class14_sub2_sub19 = ((Class14_Sub2_Sub19) Class150.aClass52_2406.get(l));
 		if (class14_sub2_sub19 != null) {
@@ -673,7 +673,7 @@ public class Class94 {
 			Class14_Sub2_Sub19 class14_sub2_sub19 = null;
 			return class14_sub2_sub19;
 		}
-		Class14_Sub2_Sub19 class14_sub2_sub19 = ((Class14_Sub2_Sub19) Class150.aClass52_2406.get((long) i_52_));
+		Class14_Sub2_Sub19 class14_sub2_sub19 = ((Class14_Sub2_Sub19) Class150.aClass52_2406.get(i_52_));
 		if (i < 30) {
 			Class14_Sub2_Sub19 class14_sub2_sub19_53_ = null;
 			return class14_sub2_sub19_53_;
@@ -686,7 +686,7 @@ public class Class94 {
 		if (class14_sub2_sub19 == null)
 			aBoolean1554 = true;
 		else
-			Class150.aClass52_2406.put(class14_sub2_sub19, (long) i_52_);
+			Class150.aClass52_2406.put(class14_sub2_sub19, i_52_);
 		Class14_Sub2_Sub19 class14_sub2_sub19_55_ = class14_sub2_sub19;
 		return class14_sub2_sub19_55_;
 	}
@@ -711,7 +711,7 @@ public class Class94 {
 			return class133_sub7;
 		}
 		if (i_57_ == 1) {
-			Class133_Sub7 class133_sub7 = ((Class133_Sub7) (JunkTex.aClass52_4588.get((long) (i_58_ + (i_57_ << 48)))));
+			Class133_Sub7 class133_sub7 = ((Class133_Sub7) (JunkTex.aClass52_4588.get(i_58_ + (i_57_ << 48))));
 			if (class133_sub7 == null) {
 				Class133_Sub2 class133_sub2 = Class133_Sub2.method1824(JunkTex.aClass9_4486, i_58_, 0);
 				if (class133_sub2 == null) {
@@ -720,7 +720,7 @@ public class Class94 {
 					return class133_sub7_59_;
 				}
 				class133_sub7 = class133_sub2.method1827(64, 768, -50, -10, -50);
-				JunkTex.aClass52_4588.put(class133_sub7, (long) ((i_57_ << 48) + i_58_));
+				JunkTex.aClass52_4588.put(class133_sub7, (i_57_ << 48) + i_58_);
 			}
 			if (animation != null)
 				class133_sub7 = animation.method1167(i_56_, 65535, class133_sub7);
